+++ examples/spot/market/ticker_price.py
@@ -7,5 +7,8 @@ from binance.lib.utils import config_logging
 config_logging(logging, logging.DEBUG)
 
 spot_client = Client(base_url="https://testnet.binance.vision")
-
-logging.info(spot_client.ticker_price("BTCUSDT"))
+symbols = ["BTCUSDT", "BNBUSDT"]
+symbol = "BNBUSDT"
+logging.info(spot_client.ticker_price())
+logging.info(spot_client.ticker_price(symbol=symbol))
+logging.info(spot_client.ticker_price(symbols=symbols))